pyproject.toml
[project]
name = "openai-vivalehack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Afnan Khan", email = "afnan.khan@davia.ai" }
]
requires-python = ">=3.12"
dependencies = [
    "davia>=0.1.13",
    "matplotlib>=3.10.3",
    "openai-agents>=0.0.17",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "replicate>=1.0.7",
]

[project.scripts]
openai-vivalehack = "collaborative-photoshop.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.11.13",
]
